requirements.txt
# Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[job-queue]

# YouTube下载
yt-dlp==2023.12.30
pytube==15.0.0

# Spotify API
spotipy==2.23.0

# 数据库
sqlalchemy==2.0.23
aiosqlite==0.19.0

# 工具库
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1

# 音频处理
pydub==0.25.1
mutagen==1.47.0

# 日志
loguru==0.7.2
